God boss somewhat working, still lots of problems
The boss is spawning again, but when for some reason whatever we committed at the end of class got messed up.

The final projectile that kills the enemy does not get destroyed for some reason, and especially with the boss where all the projectiles start piling up.

the Boss speed and health modifiers that are adjustable in the inspector dont seem to take effect, so thats another issue

not good